Move parsing of Instance attributes to GET request
The parsing of and Instance `data` and `name` attributes is now done in
the `GET` method requests. Beforehand, the attributes were taken from
the database and sent to the client as strings, the conversion to JSON
object was done only when necessary. Now the conversion is done
immediatly, since it best reflects the Instance object.