Make circle around conceptually selected node
Correction to previous commit. The selected node is now the one present
in the `data` field of the Event object, if it exists, or the `node`
field of the Event object otherwise. Previously, the `data` field was
not checked, and the selected node was taken directly from the `node`
field of the Event object.