Compute balanceness of node and show it on tree view
The balanceness of a node (average height divided by the maximum height
of children nodes) is a value going from 0 (bad) to 1 (good). It is
displayed in the data table. Moreover, a red-to-green halo surrounds the
OR nodes, the color depending on the degree of balanceness (0 is red,
while 1 is green).