Add websocket for CNF clauses/learnts transmission
Previously CNFs were sent synchronously from the GUI server to the
client. This resulted in blocking the server for a short time, and bad
shut down in case of errors on the SMTS side. Now the CNF transmission
is done asynchronously, meaning that the GUI server can be sut down in
case of errors, and it doesn't block while waiting for data.