[Skip Ci] Fix the stack_overflow on parsing QF_LIA instances in thread-architecure by preprocessing base instance
SMST's client process will not be initialised with a pool of threads, rather will start with only one thread (main), as soon as it receive initial event which will be a solve command to start solving the instance, it will not queue the event , instead it start to parse the base instance and after it has parsed the base instance successfully then it will set up the whole thread-architecture. in this way the client/ParseTerm will not share the stack with other threads on parsing base instance.
This will not solve the problem of memory exhaustion, or stucking, as
long as openSMT sequential can parse instance, SMTS too.